requirements.txt
loguru==0.7.2
pydantic==2.18.2
pydantic-settings==2.2.1
selenium==4.20.0
rpaframework==28.5.1
pandas==2.2.2
openpyxl==3.1.2
